checkpoint: Research UX through Task 1047-R2
Commit the cumulative Leonardo V2 source tree through:

- Research Suite restoration and application integration;
- Financial Tools and Artifact workflow validation;
- Task 1046-R4 table alignment and Financial Tools geometry;
- Task 1047-R2 compact Study labels;
- Studies Manager parameters, sources, origin, saved state, and pane details;
- stable UTC, P&T, and Strategy numbering;
- canonical oscillator guide-detail ordering.

Task 1047-R2 remains pending native visual acceptance.

Reference package:
Leo_V2_1047R2_PATCH_20260730_1726.zip

Reference SHA-256:
5989f980ac090b9f45a23d00e280127a0b766760067785bdc948d2a012c210a8

+++ src/leonardo/financial_tools/_calculation/oscillators.py
@@ -53,7 +53,13 @@ def _calculate_tdirsi(
     data: pd.DataFrame, parameters: Mapping[str, object], bindings: Mapping[str, object]
 ) -> tuple[tuple[pd.Series, ...], Mapping[str, object]]:
     close = _numeric_series(data, "close", context="tdirsi")
-    rsi = _rsi_values(close, int(parameters["period"]))
+    # OLD TDI consumes the finalized public RSI line after float32
+    # conversion, then promotes it for subsequent rolling calculations.
+    # This explicit quantization boundary preserves donor-exact behavior.
+    rsi = _rsi_values(
+        close,
+        int(parameters["period"]),
+    ).astype("float32").astype("float64")
     band_length = int(parameters["band_length"])
     basis = rsi.rolling(band_length, min_periods=band_length).mean()
     deviation = rsi.rolling(band_length, min_periods=band_length).std(ddof=0)
